pyproject.toml
[tool.poetry]
name = "hi_fastapi"
version = "0.1.0"
description = "Hello FastAPI"
authors = ["Tuan Nguyen Vinh <tuannv562@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.62.0"
pydantic = {extras = ["email"], version = "^1.7.3"}
alembic = "^1.4.3"
SQLAlchemy = "^1.3.20"
psycopg2 = "^2.8.6"
uvicorn = {extras = ["standard"], version = "^0.13.3"}
passlib = "^1.7.4"
python-jose = "^3.2.0"
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
black = "^20.8b1"
rope = "^0.18.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
